move HTCondorJobSubmitter to its own file
